add tests for scripts remove logs
