+++ asapdiscovery-dataviz/pyproject.toml
@@ -34,6 +34,7 @@ display-aligned-2d-molecules = "asapdiscovery.dataviz.scripts.display_aligned_2d
 split-sdf = "asapdiscovery.dataviz.scripts.split_sdf:main"
 traj-to-viz = "asapdiscovery.dataviz.scripts.traj_to_viz:main"
 pose-to-viz = "asapdiscovery.dataviz.scripts.pose_to_viz:main"
+pose-to-viz-pymol = "asapdiscovery.dataviz.scripts.pose_to_viz_pymol:main"
 
 [tool.setuptools.packages.find]
 where = ["."]
